Crate hotshot_contract_adapter

Cross-domain (between Solidity and Rust) utilities for type conversion and testing

Modules

bindings 🔒

This module contains the sol! generated bindings for solidity contracts. This is autogenerated code. Do not manually edit these files. These files may be overwritten by the codegen system at any time.

evm

jellyfish

Helpers for connecting types between Jellyfish and Solidity. Usually used during differential testing (via FFI).

light_client

Helpers and test mocks for Light Client logic

sol_types

Solidity types for interacting with contracts Re-export types that are used, sometimes renamed to avoid collision.

stake_table

Functions

field_to_u256

convert a field element to U256, panic if field size is larger than 256 bit

u256_to_field

convert U256 to a field (mod order)