hotshot_contract_adapter::bindings::timelock::Timelock

Enum TimelockCalls

pub enum TimelockCalls {

    CANCELLER_ROLE(CANCELLER_ROLECall),
    DEFAULT_ADMIN_ROLE(DEFAULT_ADMIN_ROLECall),
    EXECUTOR_ROLE(EXECUTOR_ROLECall),
    PROPOSER_ROLE(PROPOSER_ROLECall),
    cancel(cancelCall),
    execute(executeCall),
    executeBatch(executeBatchCall),
    getMinDelay(getMinDelayCall),
    getOperationState(getOperationStateCall),
    getRoleAdmin(getRoleAdminCall),
    getTimestamp(getTimestampCall),
    grantRole(grantRoleCall),
    hasRole(hasRoleCall),
    hashOperation(hashOperationCall),
    hashOperationBatch(hashOperationBatchCall),
    isOperation(isOperationCall),
    isOperationDone(isOperationDoneCall),
    isOperationPending(isOperationPendingCall),
    isOperationReady(isOperationReadyCall),
    onERC1155BatchReceived(onERC1155BatchReceivedCall),
    onERC1155Received(onERC1155ReceivedCall),
    onERC721Received(onERC721ReceivedCall),
    renounceRole(renounceRoleCall),
    revokeRole(revokeRoleCall),
    schedule(scheduleCall),
    scheduleBatch(scheduleBatchCall),
    supportsInterface(supportsInterfaceCall),
    updateDelay(updateDelayCall),

}

Container for all the Timelock function calls.

Variants

CANCELLER_ROLE(CANCELLER_ROLECall)

DEFAULT_ADMIN_ROLE(DEFAULT_ADMIN_ROLECall)

EXECUTOR_ROLE(EXECUTOR_ROLECall)

PROPOSER_ROLE(PROPOSER_ROLECall)

cancel(cancelCall)

execute(executeCall)

executeBatch(executeBatchCall)

getMinDelay(getMinDelayCall)

getOperationState(getOperationStateCall)

getRoleAdmin(getRoleAdminCall)

getTimestamp(getTimestampCall)

grantRole(grantRoleCall)

hasRole(hasRoleCall)

hashOperation(hashOperationCall)

hashOperationBatch(hashOperationBatchCall)

isOperation(isOperationCall)

isOperationDone(isOperationDoneCall)

isOperationPending(isOperationPendingCall)

isOperationReady(isOperationReadyCall)

onERC1155BatchReceived(onERC1155BatchReceivedCall)

onERC1155Received(onERC1155ReceivedCall)

onERC721Received(onERC721ReceivedCall)

renounceRole(renounceRoleCall)

revokeRole(revokeRoleCall)

schedule(scheduleCall)

scheduleBatch(scheduleBatchCall)

supportsInterface(supportsInterfaceCall)

updateDelay(updateDelayCall)

Implementations

impl TimelockCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.