hotshot_contract_adapter::sol_types::EspToken

Function deploy

pub fn deploy<T: Transport + Clone, P: Provider<T, N>, N: Network>(
    provider: P,
) -> impl Future<Output = Result<EspTokenInstance<T, P, N>>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.