hotshot_contract_adapter::sol_types

Module RewardClaim

What to re-export, what to hide?

export contract struct itself, but try to avoid export instance type (instead, use ::new() to get a handle)

avoid exporting xxCall and xxReturn types, they usually can be converted/transmuted from existing struct

Event types should be exported

structs should be exported and renamed with xxSol suffix to avoid confusion with other rust types

see module doc for more explanation on types duplication issue in alloy

Generated by the following Solidity interface…

interface RewardClaim {
    error AccessControlBadConfirmation();
    error AccessControlUnauthorizedAccount(address account, bytes32 neededRole);
    error AddressEmptyCode(address target);
    error AlreadyClaimed();
    error DailyLimitExceeded();
    error DailyLimitTooHigh();
    error ERC1967InvalidImplementation(address implementation);
    error ERC1967NonPayable();
    error EnforcedPause();
    error ExpectedPause();
    error FailedInnerCall();
    error InvalidAuthRoot();
    error InvalidInitialization();
    error InvalidRewardAmount();
    error NoChangeRequired();
    error NotInitializing();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error ReentrancyGuardReentrantCall();
    error UUPSUnauthorizedCallContext();
    error UUPSUnsupportedProxiableUUID(bytes32 slot);
    error ZeroDailyLimit();
    error ZeroLightClientAddress();
    error ZeroPauserAddress();
    error ZeroTokenAddress();
    error ZeroTotalSupply();

    event DailyLimitUpdated(uint256 oldLimit, uint256 newLimit);
    event Initialized(uint64 version);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event Paused(address account);
    event RewardsClaimed(address indexed user, uint256 amount);
    event RoleAdminChanged(bytes32 indexed role, bytes32 indexed previousAdminRole, bytes32 indexed newAdminRole);
    event RoleGranted(bytes32 indexed role, address indexed account, address indexed sender);
    event RoleRevoked(bytes32 indexed role, address indexed account, address indexed sender);
    event Unpaused(address account);
    event Upgrade(address implementation);
    event Upgraded(address indexed implementation);

    constructor();

    function DEFAULT_ADMIN_ROLE() external view returns (bytes32);
    function MAX_DAILY_LIMIT_BASIS_POINTS() external view returns (uint256);
    function PAUSER_ROLE() external view returns (bytes32);
    function UPGRADE_INTERFACE_VERSION() external view returns (string memory);
    function claimRewards(uint256 lifetimeRewards, bytes memory authData) external;
    function claimedRewards(address claimer) external view returns (uint256 claimed);
    function dailyLimitWei() external view returns (uint256);
    function espToken() external view returns (address);
    function getRoleAdmin(bytes32 role) external view returns (bytes32);
    function getVersion() external pure returns (uint8 majorVersion, uint8 minorVersion, uint8 patchVersion);
    function grantRole(bytes32 role, address account) external;
    function hasRole(bytes32 role, address account) external view returns (bool);
    function initialize(address _owner, address _espToken, address _lightClient, address _pauser) external;
    function lastSetDailyLimitBasisPoints() external view returns (uint256);
    function lightClient() external view returns (address);
    function owner() external view returns (address);
    function pause() external;
    function paused() external view returns (bool);
    function proxiableUUID() external view returns (bytes32);
    function renounceOwnership() external;
    function renounceRole(bytes32 role, address callerConfirmation) external;
    function revokeRole(bytes32 role, address account) external;
    function setDailyLimit(uint256 basisPoints) external;
    function supportsInterface(bytes4 interfaceId) external view returns (bool);
    function transferOwnership(address newOwner) external;
    function unpause() external;
    function upgradeToAndCall(address newImplementation, bytes memory data) external payable;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "DEFAULT_ADMIN_ROLE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_DAILY_LIMIT_BASIS_POINTS",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "PAUSER_ROLE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "UPGRADE_INTERFACE_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "claimRewards",
    "inputs": [
      {
        "name": "lifetimeRewards",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "authData",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "claimedRewards",
    "inputs": [
      {
        "name": "claimer",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "claimed",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "dailyLimitWei",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "espToken",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract EspTokenV2"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getRoleAdmin",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getVersion",
    "inputs": [],
    "outputs": [
      {
        "name": "majorVersion",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "minorVersion",
        "type": "uint8",
        "internalType": "uint8"
      },
      {
        "name": "patchVersion",
        "type": "uint8",
        "internalType": "uint8"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "grantRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "hasRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "_owner",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_espToken",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_lightClient",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_pauser",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "lastSetDailyLimitBasisPoints",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "lightClient",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract LightClientV3"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pause",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "paused",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "proxiableUUID",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "renounceRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "callerConfirmation",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revokeRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setDailyLimit",
    "inputs": [
      {
        "name": "basisPoints",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "supportsInterface",
    "inputs": [
      {
        "name": "interfaceId",
        "type": "bytes4",
        "internalType": "bytes4"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "unpause",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "upgradeToAndCall",
    "inputs": [
      {
        "name": "newImplementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "event",
    "name": "DailyLimitUpdated",
    "inputs": [
      {
        "name": "oldLimit",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "newLimit",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Paused",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RewardsClaimed",
    "inputs": [
      {
        "name": "user",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RoleAdminChanged",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "previousAdminRole",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "newAdminRole",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RoleGranted",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RoleRevoked",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Unpaused",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgrade",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgraded",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AccessControlBadConfirmation",
    "inputs": []
  },
  {
    "type": "error",
    "name": "AccessControlUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "neededRole",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AlreadyClaimed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DailyLimitExceeded",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DailyLimitTooHigh",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ERC1967InvalidImplementation",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967NonPayable",
    "inputs": []
  },
  {
    "type": "error",
    "name": "EnforcedPause",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ExpectedPause",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedInnerCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidAuthRoot",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidRewardAmount",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NoChangeRequired",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ReentrancyGuardReentrantCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnauthorizedCallContext",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnsupportedProxiableUUID",
    "inputs": [
      {
        "name": "slot",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  },
  {
    "type": "error",
    "name": "ZeroDailyLimit",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroLightClientAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroPauserAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroTokenAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroTotalSupply",
    "inputs": []
  }
]

Structs

AccessControlBadConfirmation

Custom error with signature AccessControlBadConfirmation() and selector 0x6697b232.

AccessControlUnauthorizedAccount

Custom error with signature AccessControlUnauthorizedAccount(address,bytes32) and selector 0xe2517d3f.

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AlreadyClaimed

Custom error with signature AlreadyClaimed() and selector 0x646cf558.

DEFAULT_ADMIN_ROLECall

Function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf.

DEFAULT_ADMIN_ROLEReturn

Container type for the return parameters of the DEFAULT_ADMIN_ROLE() function.

DailyLimitExceeded

Custom error with signature DailyLimitExceeded() and selector 0x194bd314.

DailyLimitTooHigh

Custom error with signature DailyLimitTooHigh() and selector 0xc5d2ae53.

DailyLimitUpdated

Event with signature DailyLimitUpdated(uint256,uint256) and selector 0x207c4cbdf55ec315a13f0d5e047732ec5d947da056e706593aa509909941cedf.

ERC1967InvalidImplementation

Custom error with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3.

ERC1967NonPayable

Custom error with signature ERC1967NonPayable() and selector 0xb398979f.

EnforcedPause

Custom error with signature EnforcedPause() and selector 0xd93c0665.

ExpectedPause

Custom error with signature ExpectedPause() and selector 0x8dfc202b.

FailedInnerCall

Custom error with signature FailedInnerCall() and selector 0x1425ea42.

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2.

InvalidAuthRoot

Custom error with signature InvalidAuthRoot() and selector 0x328b8878.

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9.

InvalidRewardAmount

Custom error with signature InvalidRewardAmount() and selector 0x38539865.

MAX_DAILY_LIMIT_BASIS_POINTSCall

Function with signature MAX_DAILY_LIMIT_BASIS_POINTS() and selector 0xcf21395b.

MAX_DAILY_LIMIT_BASIS_POINTSReturn

Container type for the return parameters of the MAX_DAILY_LIMIT_BASIS_POINTS() function.

NoChangeRequired

Custom error with signature NoChangeRequired() and selector 0xa863aec9.

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8.

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7.

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

PAUSER_ROLECall

Function with signature PAUSER_ROLE() and selector 0xe63ab1e9.

PAUSER_ROLEReturn

Container type for the return parameters of the PAUSER_ROLE() function.

Paused

Event with signature Paused(address) and selector 0x62e78cea01bee320cd4e420270b5ea74000d11b0c9f74754ebdbfc544b05a258.

ReentrancyGuardReentrantCall

Custom error with signature ReentrancyGuardReentrantCall() and selector 0x3ee5aeb5.

RewardClaimInstance

A RewardClaim instance.

RewardsClaimed

Event with signature RewardsClaimed(address,uint256) and selector 0xfc30cddea38e2bf4d6ea7d3f9ed3b6ad7f176419f4963bd81318067a4aee73fe.

RoleAdminChanged

Event with signature RoleAdminChanged(bytes32,bytes32,bytes32) and selector 0xbd79b86ffe0ab8e8776151514217cd7cacd52c909f66475c3af44e129f0b00ff.

RoleGranted

Event with signature RoleGranted(bytes32,address,address) and selector 0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d.

RoleRevoked

Event with signature RoleRevoked(bytes32,address,address) and selector 0xf6391f5c32d9c69d2a47ea670b442974b53935d1edc7fd64eb21e047a839171b.

UPGRADE_INTERFACE_VERSIONCall

Function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc.

UPGRADE_INTERFACE_VERSIONReturn

Container type for the return parameters of the UPGRADE_INTERFACE_VERSION() function.

UUPSUnauthorizedCallContext

Custom error with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba.

UUPSUnsupportedProxiableUUID

Custom error with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4.

Unpaused

Event with signature Unpaused(address) and selector 0x5db9ee0a495bf2e6ff9c91a7834c1ba4fdd244a5e8aa4e537bd38aeae4b073aa.

Upgrade

Event with signature Upgrade(address) and selector 0xf78721226efe9a1bb678189a16d1554928b9f2192e2cb93eeda83b79fa40007d.

Upgraded

Event with signature Upgraded(address) and selector 0xbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b.

ZeroDailyLimit

Custom error with signature ZeroDailyLimit() and selector 0x9b11a8bd.

ZeroLightClientAddress

Custom error with signature ZeroLightClientAddress() and selector 0xfcabddbd.

ZeroPauserAddress

Custom error with signature ZeroPauserAddress() and selector 0xa74995ab.

ZeroTokenAddress

Custom error with signature ZeroTokenAddress() and selector 0x6b093aad.

ZeroTotalSupply

Custom error with signature ZeroTotalSupply() and selector 0x48b50023.

claimRewardsCall

Function with signature claimRewards(uint256,bytes) and selector 0x549dd8c3.

claimRewardsReturn

Container type for the return parameters of the claimRewards(uint256,bytes) function.

claimedRewardsCall

Function with signature claimedRewards(address) and selector 0xbd834345.

claimedRewardsReturn

Container type for the return parameters of the claimedRewards(address) function.

constructorCall

Constructor`.

dailyLimitWeiCall

Function with signature dailyLimitWei() and selector 0x2cca74d8.

dailyLimitWeiReturn

Container type for the return parameters of the dailyLimitWei() function.

espTokenCall

Function with signature espToken() and selector 0xf092e13a.

espTokenReturn

Container type for the return parameters of the espToken() function.

getRoleAdminCall

Function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3.

getRoleAdminReturn

Container type for the return parameters of the getRoleAdmin(bytes32) function.

getVersionCall

Function with signature getVersion() and selector 0x0d8e6e2c.

getVersionReturn

Container type for the return parameters of the getVersion() function.

grantRoleCall

Function with signature grantRole(bytes32,address) and selector 0x2f2ff15d.

grantRoleReturn

Container type for the return parameters of the grantRole(bytes32,address) function.

hasRoleCall

Function with signature hasRole(bytes32,address) and selector 0x91d14854.

hasRoleReturn

Container type for the return parameters of the hasRole(bytes32,address) function.

initializeCall

Function with signature initialize(address,address,address,address) and selector 0xf8c8765e.

initializeReturn

Container type for the return parameters of the initialize(address,address,address,address) function.

lastSetDailyLimitBasisPointsCall

Function with signature lastSetDailyLimitBasisPoints() and selector 0x0f6ca1d6.

lastSetDailyLimitBasisPointsReturn

Container type for the return parameters of the lastSetDailyLimitBasisPoints() function.

lightClientCall

Function with signature lightClient() and selector 0xb5700e68.

lightClientReturn

Container type for the return parameters of the lightClient() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

pauseCall

Function with signature pause() and selector 0x8456cb59.

pauseReturn

Container type for the return parameters of the pause() function.

pausedCall

Function with signature paused() and selector 0x5c975abb.

pausedReturn

Container type for the return parameters of the paused() function.

proxiableUUIDCall

Function with signature proxiableUUID() and selector 0x52d1902d.

proxiableUUIDReturn

Container type for the return parameters of the proxiableUUID() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

renounceRoleCall

Function with signature renounceRole(bytes32,address) and selector 0x36568abe.

renounceRoleReturn

Container type for the return parameters of the renounceRole(bytes32,address) function.

revokeRoleCall

Function with signature revokeRole(bytes32,address) and selector 0xd547741f.

revokeRoleReturn

Container type for the return parameters of the revokeRole(bytes32,address) function.

setDailyLimitCall

Function with signature setDailyLimit(uint256) and selector 0xb20d30a9.

setDailyLimitReturn

Container type for the return parameters of the setDailyLimit(uint256) function.

supportsInterfaceCall

Function with signature supportsInterface(bytes4) and selector 0x01ffc9a7.

supportsInterfaceReturn

Container type for the return parameters of the supportsInterface(bytes4) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

unpauseCall

Function with signature unpause() and selector 0x3f4ba83a.

unpauseReturn

Container type for the return parameters of the unpause() function.

upgradeToAndCallCall

Function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286.

upgradeToAndCallReturn

Container type for the return parameters of the upgradeToAndCall(address,bytes) function.

Enums

RewardClaimCalls

Container for all the RewardClaim function calls.

RewardClaimErrors

Container for all the RewardClaim custom errors.

RewardClaimEvents

Container for all the RewardClaim events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain RewardClaim contract instance.