hotshot_contract_adapter::sol_types::Timelock

Struct TimelockInstance

pub struct TimelockInstance<T, P, N = Ethereum> {
    address: Address,
    provider: P,
    _network_transport: PhantomData<(N, T)>,
}

A Timelock instance.

Contains type-safe methods for interacting with an on-chain instance of the
[`Timelock`](self) contract located at a given `address`, using a given
provider `P`.

If the contract bytecode is available (see the [`sol!`](alloy_sol_types::sol!)
documentation on how to provide it), the `deploy` and `deploy_builder` methods can
be used to deploy a new instance of the contract.

See the [module-level documentation](self) for all the available methods.

Fields
address: Addressprovider: P_network_transport: PhantomData<(N, T)>
Implementations

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> TimelockInstance<T, P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain Timelock contract instance.

    See the [wrapper's documentation](`TimelockInstance`) for more details.

pub async fn deploy( provider: P, minDelay: U256, proposers: Vec<Address>, executors: Vec<Address>, admin: Address, ) -> Result<TimelockInstance<T, P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

    Returns a new instance of the contract, if the deployment was successful.

    For more fine-grained control over the deployment process, use [`deploy_builder`] instead.

pub fn deploy_builder( provider: P, minDelay: U256, proposers: Vec<Address>, executors: Vec<Address>, admin: Address, ) -> RawCallBuilder<T, P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

    This is a simple wrapper around creating a `RawCallBuilder` with the data set to
    the bytecode concatenated with the constructor's ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<T, P: Clone, N> TimelockInstance<T, &P, N>

pub fn with_cloned_provider(self) -> TimelockInstance<T, P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> TimelockInstance<T, P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<T, &P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn CANCELLER_ROLE(&self) -> SolCallBuilder<T, &P, CANCELLER_ROLECall, N>

Creates a new call builder for the [CANCELLER_ROLE] function.

pub fn DEFAULT_ADMIN_ROLE( &self, ) -> SolCallBuilder<T, &P, DEFAULT_ADMIN_ROLECall, N>

Creates a new call builder for the [DEFAULT_ADMIN_ROLE] function.

pub fn EXECUTOR_ROLE(&self) -> SolCallBuilder<T, &P, EXECUTOR_ROLECall, N>

Creates a new call builder for the [EXECUTOR_ROLE] function.

pub fn PROPOSER_ROLE(&self) -> SolCallBuilder<T, &P, PROPOSER_ROLECall, N>

Creates a new call builder for the [PROPOSER_ROLE] function.

pub fn cancel(&self, id: FixedBytes<32>) -> SolCallBuilder<T, &P, cancelCall, N>

Creates a new call builder for the [cancel] function.

pub fn execute( &self, target: Address, value: U256, payload: Bytes, predecessor: FixedBytes<32>, salt: FixedBytes<32>, ) -> SolCallBuilder<T, &P, executeCall, N>

Creates a new call builder for the [execute] function.

pub fn executeBatch( &self, targets: Vec<Address>, values: Vec<U256>, payloads: Vec<Bytes>, predecessor: FixedBytes<32>, salt: FixedBytes<32>, ) -> SolCallBuilder<T, &P, executeBatchCall, N>

Creates a new call builder for the [executeBatch] function.

pub fn getMinDelay(&self) -> SolCallBuilder<T, &P, getMinDelayCall, N>

Creates a new call builder for the [getMinDelay] function.

pub fn getOperationState( &self, id: FixedBytes<32>, ) -> SolCallBuilder<T, &P, getOperationStateCall, N>

Creates a new call builder for the [getOperationState] function.

pub fn getRoleAdmin( &self, role: FixedBytes<32>, ) -> SolCallBuilder<T, &P, getRoleAdminCall, N>

Creates a new call builder for the [getRoleAdmin] function.

pub fn getTimestamp( &self, id: FixedBytes<32>, ) -> SolCallBuilder<T, &P, getTimestampCall, N>

Creates a new call builder for the [getTimestamp] function.

pub fn grantRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<T, &P, grantRoleCall, N>

Creates a new call builder for the [grantRole] function.

pub fn hasRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<T, &P, hasRoleCall, N>

Creates a new call builder for the [hasRole] function.

pub fn hashOperation( &self, target: Address, value: U256, data: Bytes, predecessor: FixedBytes<32>, salt: FixedBytes<32>, ) -> SolCallBuilder<T, &P, hashOperationCall, N>

Creates a new call builder for the [hashOperation] function.

pub fn hashOperationBatch( &self, targets: Vec<Address>, values: Vec<U256>, payloads: Vec<Bytes>, predecessor: FixedBytes<32>, salt: FixedBytes<32>, ) -> SolCallBuilder<T, &P, hashOperationBatchCall, N>

Creates a new call builder for the [hashOperationBatch] function.

pub fn isOperation( &self, id: FixedBytes<32>, ) -> SolCallBuilder<T, &P, isOperationCall, N>

Creates a new call builder for the [isOperation] function.

pub fn isOperationDone( &self, id: FixedBytes<32>, ) -> SolCallBuilder<T, &P, isOperationDoneCall, N>

Creates a new call builder for the [isOperationDone] function.

pub fn isOperationPending( &self, id: FixedBytes<32>, ) -> SolCallBuilder<T, &P, isOperationPendingCall, N>

Creates a new call builder for the [isOperationPending] function.

pub fn isOperationReady( &self, id: FixedBytes<32>, ) -> SolCallBuilder<T, &P, isOperationReadyCall, N>

Creates a new call builder for the [isOperationReady] function.

pub fn onERC1155BatchReceived( &self, _0: Address, _1: Address, _2: Vec<U256>, _3: Vec<U256>, _4: Bytes, ) -> SolCallBuilder<T, &P, onERC1155BatchReceivedCall, N>

Creates a new call builder for the [onERC1155BatchReceived] function.

pub fn onERC1155Received( &self, _0: Address, _1: Address, _2: U256, _3: U256, _4: Bytes, ) -> SolCallBuilder<T, &P, onERC1155ReceivedCall, N>

Creates a new call builder for the [onERC1155Received] function.

pub fn onERC721Received( &self, _0: Address, _1: Address, _2: U256, _3: Bytes, ) -> SolCallBuilder<T, &P, onERC721ReceivedCall, N>

Creates a new call builder for the [onERC721Received] function.

pub fn renounceRole( &self, role: FixedBytes<32>, callerConfirmation: Address, ) -> SolCallBuilder<T, &P, renounceRoleCall, N>

Creates a new call builder for the [renounceRole] function.

pub fn revokeRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<T, &P, revokeRoleCall, N>

Creates a new call builder for the [revokeRole] function.

pub fn schedule( &self, target: Address, value: U256, data: Bytes, predecessor: FixedBytes<32>, salt: FixedBytes<32>, delay: U256, ) -> SolCallBuilder<T, &P, scheduleCall, N>

Creates a new call builder for the [schedule] function.

pub fn scheduleBatch( &self, targets: Vec<Address>, values: Vec<U256>, payloads: Vec<Bytes>, predecessor: FixedBytes<32>, salt: FixedBytes<32>, delay: U256, ) -> SolCallBuilder<T, &P, scheduleBatchCall, N>

Creates a new call builder for the [scheduleBatch] function.

pub fn supportsInterface( &self, interfaceId: FixedBytes<4>, ) -> SolCallBuilder<T, &P, supportsInterfaceCall, N>

Creates a new call builder for the [supportsInterface] function.

pub fn updateDelay( &self, newDelay: U256, ) -> SolCallBuilder<T, &P, updateDelayCall, N>

Creates a new call builder for the [updateDelay] function.

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> TimelockInstance<T, P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<T, &P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn CallExecuted_filter(&self) -> Event<T, &P, CallExecuted, N>

Creates a new event filter for the CallExecuted event.

pub fn CallSalt_filter(&self) -> Event<T, &P, CallSalt, N>

Creates a new event filter for the CallSalt event.

pub fn CallScheduled_filter(&self) -> Event<T, &P, CallScheduled, N>

Creates a new event filter for the CallScheduled event.

pub fn Cancelled_filter(&self) -> Event<T, &P, Cancelled, N>

Creates a new event filter for the Cancelled event.

pub fn MinDelayChange_filter(&self) -> Event<T, &P, MinDelayChange, N>

Creates a new event filter for the MinDelayChange event.

pub fn RoleAdminChanged_filter(&self) -> Event<T, &P, RoleAdminChanged, N>

Creates a new event filter for the RoleAdminChanged event.

pub fn RoleGranted_filter(&self) -> Event<T, &P, RoleGranted, N>

Creates a new event filter for the RoleGranted event.

pub fn RoleRevoked_filter(&self) -> Event<T, &P, RoleRevoked, N>

Creates a new event filter for the RoleRevoked event.