hotshot_contract_adapter::sol_types::reward_claim_prototype_mock

Module RewardClaimPrototypeMock

Generated by the following Solidity interface…

interface RewardClaimPrototypeMock {
    function verifyRewardClaim(bytes32 root, address account, uint256 amount, bytes32[160] memory proof) external pure returns (bool);
    function verifyRewardClaimAuthData(bytes32 root, address account, uint256 amount, bytes memory authData) external view returns (bool);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "function",
    "name": "verifyRewardClaim",
    "inputs": [
      {
        "name": "root",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "proof",
        "type": "bytes32[160]",
        "internalType": "bytes32[160]"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "verifyRewardClaimAuthData",
    "inputs": [
      {
        "name": "root",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "authData",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  }
]

Structs

RewardClaimPrototypeMockInstance

A RewardClaimPrototypeMock instance.

verifyRewardClaimAuthDataCall

Function with signature verifyRewardClaimAuthData(bytes32,address,uint256,bytes) and selector 0xeb2dbe08.

verifyRewardClaimAuthDataReturn

Container type for the return parameters of the verifyRewardClaimAuthData(bytes32,address,uint256,bytes) function.

verifyRewardClaimCall

Function with signature verifyRewardClaim(bytes32,address,uint256,bytes32[160]) and selector 0x24bb9778.

verifyRewardClaimReturn

Container type for the return parameters of the verifyRewardClaim(bytes32,address,uint256,bytes32[160]) function.

Enums

RewardClaimPrototypeMockCalls

Container for all the RewardClaimPrototypeMock function calls.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain RewardClaimPrototypeMock contract instance.