hotshot_contract_adapter::sol_types::stake_table_v2::StakeTableV2

Struct StakeTableV2Instance

pub struct StakeTableV2Instance<P, N = Ethereum> {
    address: Address,
    provider: P,
    _network: PhantomData<N>,
}

A StakeTableV2 instance.

Contains type-safe methods for interacting with an on-chain instance of the StakeTableV2 contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Fields
address: Addressprovider: P_network: PhantomData<N>
Implementations

impl<P: Provider<N>, N: Network> StakeTableV2Instance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain StakeTableV2 contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy(provider: P) -> Result<StakeTableV2Instance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder(provider: P) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> StakeTableV2Instance<&P, N>

pub fn with_cloned_provider(self) -> StakeTableV2Instance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> StakeTableV2Instance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn DEFAULT_ADMIN_ROLE( &self, ) -> SolCallBuilder<&P, DEFAULT_ADMIN_ROLECall, N>

Creates a new call builder for the [DEFAULT_ADMIN_ROLE] function.

pub fn PAUSER_ROLE(&self) -> SolCallBuilder<&P, PAUSER_ROLECall, N>

Creates a new call builder for the [PAUSER_ROLE] function.

pub fn UPGRADE_INTERFACE_VERSION( &self, ) -> SolCallBuilder<&P, UPGRADE_INTERFACE_VERSIONCall, N>

Creates a new call builder for the [UPGRADE_INTERFACE_VERSION] function.

pub fn _hashBlsKey( &self, blsVK: <G2Point as SolType>::RustType, ) -> SolCallBuilder<&P, _hashBlsKeyCall, N>

Creates a new call builder for the [_hashBlsKey] function.

pub fn blsKeys( &self, blsKeyHash: FixedBytes<32>, ) -> SolCallBuilder<&P, blsKeysCall, N>

Creates a new call builder for the [blsKeys] function.

pub fn claimValidatorExit( &self, validator: Address, ) -> SolCallBuilder<&P, claimValidatorExitCall, N>

Creates a new call builder for the [claimValidatorExit] function.

pub fn claimWithdrawal( &self, validator: Address, ) -> SolCallBuilder<&P, claimWithdrawalCall, N>

Creates a new call builder for the [claimWithdrawal] function.

pub fn commissionTracking( &self, validator: Address, ) -> SolCallBuilder<&P, commissionTrackingCall, N>

Creates a new call builder for the [commissionTracking] function.

pub fn delegate( &self, validator: Address, amount: U256, ) -> SolCallBuilder<&P, delegateCall, N>

Creates a new call builder for the [delegate] function.

pub fn delegations( &self, validator: Address, delegator: Address, ) -> SolCallBuilder<&P, delegationsCall, N>

Creates a new call builder for the [delegations] function.

pub fn deregisterValidator( &self, ) -> SolCallBuilder<&P, deregisterValidatorCall, N>

Creates a new call builder for the [deregisterValidator] function.

pub fn exitEscrowPeriod(&self) -> SolCallBuilder<&P, exitEscrowPeriodCall, N>

Creates a new call builder for the [exitEscrowPeriod] function.

pub fn getRoleAdmin( &self, role: FixedBytes<32>, ) -> SolCallBuilder<&P, getRoleAdminCall, N>

Creates a new call builder for the [getRoleAdmin] function.

pub fn getVersion(&self) -> SolCallBuilder<&P, getVersionCall, N>

Creates a new call builder for the [getVersion] function.

pub fn grantRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, grantRoleCall, N>

Creates a new call builder for the [grantRole] function.

pub fn hasRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, hasRoleCall, N>

Creates a new call builder for the [hasRole] function.

pub fn initialize( &self, _tokenAddress: Address, _lightClientAddress: Address, _exitEscrowPeriod: U256, _timelock: Address, ) -> SolCallBuilder<&P, initializeCall, N>

Creates a new call builder for the [initialize] function.

pub fn initializeV2( &self, pauser: Address, admin: Address, initialCommissions: Vec<<InitialCommission as SolType>::RustType>, ) -> SolCallBuilder<&P, initializeV2Call, N>

Creates a new call builder for the [initializeV2] function.

pub fn initializedAtBlock( &self, ) -> SolCallBuilder<&P, initializedAtBlockCall, N>

Creates a new call builder for the [initializedAtBlock] function.

pub fn lightClient(&self) -> SolCallBuilder<&P, lightClientCall, N>

Creates a new call builder for the [lightClient] function.

pub fn maxCommissionIncrease( &self, ) -> SolCallBuilder<&P, maxCommissionIncreaseCall, N>

Creates a new call builder for the [maxCommissionIncrease] function.

pub fn minCommissionIncreaseInterval( &self, ) -> SolCallBuilder<&P, minCommissionIncreaseIntervalCall, N>

Creates a new call builder for the [minCommissionIncreaseInterval] function.

pub fn owner(&self) -> SolCallBuilder<&P, ownerCall, N>

Creates a new call builder for the [owner] function.

pub fn pause(&self) -> SolCallBuilder<&P, pauseCall, N>

Creates a new call builder for the [pause] function.

pub fn paused(&self) -> SolCallBuilder<&P, pausedCall, N>

Creates a new call builder for the [paused] function.

pub fn proxiableUUID(&self) -> SolCallBuilder<&P, proxiableUUIDCall, N>

Creates a new call builder for the [proxiableUUID] function.

pub fn registerValidator( &self, _0: <G2Point as SolType>::RustType, _1: <EdOnBN254Point as SolType>::RustType, _2: <G1Point as SolType>::RustType, _3: u16, ) -> SolCallBuilder<&P, registerValidatorCall, N>

Creates a new call builder for the [registerValidator] function.

pub fn registerValidatorV2( &self, blsVK: <G2Point as SolType>::RustType, schnorrVK: <EdOnBN254Point as SolType>::RustType, blsSig: <G1Point as SolType>::RustType, schnorrSig: Bytes, commission: u16, ) -> SolCallBuilder<&P, registerValidatorV2Call, N>

Creates a new call builder for the [registerValidatorV2] function.

pub fn renounceOwnership(&self) -> SolCallBuilder<&P, renounceOwnershipCall, N>

Creates a new call builder for the [renounceOwnership] function.

pub fn renounceRole( &self, role: FixedBytes<32>, callerConfirmation: Address, ) -> SolCallBuilder<&P, renounceRoleCall, N>

Creates a new call builder for the [renounceRole] function.

pub fn revokeRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, revokeRoleCall, N>

Creates a new call builder for the [revokeRole] function.

pub fn schnorrKeys( &self, schnorrKey: FixedBytes<32>, ) -> SolCallBuilder<&P, schnorrKeysCall, N>

Creates a new call builder for the [schnorrKeys] function.

pub fn setMaxCommissionIncrease( &self, newMaxIncrease: u16, ) -> SolCallBuilder<&P, setMaxCommissionIncreaseCall, N>

Creates a new call builder for the [setMaxCommissionIncrease] function.

pub fn setMinCommissionUpdateInterval( &self, newInterval: U256, ) -> SolCallBuilder<&P, setMinCommissionUpdateIntervalCall, N>

Creates a new call builder for the [setMinCommissionUpdateInterval] function.

pub fn supportsInterface( &self, interfaceId: FixedBytes<4>, ) -> SolCallBuilder<&P, supportsInterfaceCall, N>

Creates a new call builder for the [supportsInterface] function.

pub fn token(&self) -> SolCallBuilder<&P, tokenCall, N>

Creates a new call builder for the [token] function.

pub fn transferOwnership( &self, newOwner: Address, ) -> SolCallBuilder<&P, transferOwnershipCall, N>

Creates a new call builder for the [transferOwnership] function.

pub fn undelegate( &self, validator: Address, amount: U256, ) -> SolCallBuilder<&P, undelegateCall, N>

Creates a new call builder for the [undelegate] function.

pub fn undelegations( &self, validator: Address, delegator: Address, ) -> SolCallBuilder<&P, undelegationsCall, N>

Creates a new call builder for the [undelegations] function.

pub fn unpause(&self) -> SolCallBuilder<&P, unpauseCall, N>

Creates a new call builder for the [unpause] function.

pub fn updateCommission( &self, newCommission: u16, ) -> SolCallBuilder<&P, updateCommissionCall, N>

Creates a new call builder for the [updateCommission] function.

pub fn updateConsensusKeys( &self, _0: <G2Point as SolType>::RustType, _1: <EdOnBN254Point as SolType>::RustType, _2: <G1Point as SolType>::RustType, ) -> SolCallBuilder<&P, updateConsensusKeysCall, N>

Creates a new call builder for the [updateConsensusKeys] function.

pub fn updateConsensusKeysV2( &self, blsVK: <G2Point as SolType>::RustType, schnorrVK: <EdOnBN254Point as SolType>::RustType, blsSig: <G1Point as SolType>::RustType, schnorrSig: Bytes, ) -> SolCallBuilder<&P, updateConsensusKeysV2Call, N>

Creates a new call builder for the [updateConsensusKeysV2] function.

pub fn updateExitEscrowPeriod( &self, newExitEscrowPeriod: u64, ) -> SolCallBuilder<&P, updateExitEscrowPeriodCall, N>

Creates a new call builder for the [updateExitEscrowPeriod] function.

pub fn upgradeToAndCall( &self, newImplementation: Address, data: Bytes, ) -> SolCallBuilder<&P, upgradeToAndCallCall, N>

Creates a new call builder for the [upgradeToAndCall] function.

pub fn validatorExits( &self, validator: Address, ) -> SolCallBuilder<&P, validatorExitsCall, N>

Creates a new call builder for the [validatorExits] function.

pub fn validators( &self, account: Address, ) -> SolCallBuilder<&P, validatorsCall, N>

Creates a new call builder for the [validators] function.

impl<P: Provider<N>, N: Network> StakeTableV2Instance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn CommissionUpdated_filter(&self) -> Event<&P, CommissionUpdated, N>

Creates a new event filter for the CommissionUpdated event.

pub fn ConsensusKeysUpdated_filter(&self) -> Event<&P, ConsensusKeysUpdated, N>

Creates a new event filter for the ConsensusKeysUpdated event.

pub fn ConsensusKeysUpdatedV2_filter( &self, ) -> Event<&P, ConsensusKeysUpdatedV2, N>

Creates a new event filter for the ConsensusKeysUpdatedV2 event.

pub fn Delegated_filter(&self) -> Event<&P, Delegated, N>

Creates a new event filter for the Delegated event.

pub fn ExitEscrowPeriodUpdated_filter( &self, ) -> Event<&P, ExitEscrowPeriodUpdated, N>

Creates a new event filter for the ExitEscrowPeriodUpdated event.

pub fn Initialized_filter(&self) -> Event<&P, Initialized, N>

Creates a new event filter for the Initialized event.

pub fn MaxCommissionIncreaseUpdated_filter( &self, ) -> Event<&P, MaxCommissionIncreaseUpdated, N>

Creates a new event filter for the MaxCommissionIncreaseUpdated event.

pub fn MinCommissionUpdateIntervalUpdated_filter( &self, ) -> Event<&P, MinCommissionUpdateIntervalUpdated, N>

Creates a new event filter for the MinCommissionUpdateIntervalUpdated event.

pub fn OwnershipTransferred_filter(&self) -> Event<&P, OwnershipTransferred, N>

Creates a new event filter for the OwnershipTransferred event.

pub fn Paused_filter(&self) -> Event<&P, Paused, N>

Creates a new event filter for the Paused event.

pub fn RoleAdminChanged_filter(&self) -> Event<&P, RoleAdminChanged, N>

Creates a new event filter for the RoleAdminChanged event.

pub fn RoleGranted_filter(&self) -> Event<&P, RoleGranted, N>

Creates a new event filter for the RoleGranted event.

pub fn RoleRevoked_filter(&self) -> Event<&P, RoleRevoked, N>

Creates a new event filter for the RoleRevoked event.

pub fn Undelegated_filter(&self) -> Event<&P, Undelegated, N>

Creates a new event filter for the Undelegated event.

pub fn Unpaused_filter(&self) -> Event<&P, Unpaused, N>

Creates a new event filter for the Unpaused event.

pub fn Upgraded_filter(&self) -> Event<&P, Upgraded, N>

Creates a new event filter for the Upgraded event.

pub fn ValidatorExit_filter(&self) -> Event<&P, ValidatorExit, N>

Creates a new event filter for the ValidatorExit event.

pub fn ValidatorRegistered_filter(&self) -> Event<&P, ValidatorRegistered, N>

Creates a new event filter for the ValidatorRegistered event.

pub fn ValidatorRegisteredV2_filter( &self, ) -> Event<&P, ValidatorRegisteredV2, N>

Creates a new event filter for the ValidatorRegisteredV2 event.

pub fn Withdrawal_filter(&self) -> Event<&P, Withdrawal, N>

Creates a new event filter for the Withdrawal event.